requirements.txt
cython
matplotlib
opencv-python
numpy==1.19.5
tqdm==4.55.1
mmcv-full==1.2.4
mmdet==2.8.0
